Bump polaris version from 1.1 to 1.2
We have Polaris running as part of regression tests. Polaris 1.2 is released, and let's try to be up-to-date with Polaris.

There is one relatively important change that we are impacted by:

```
With version 1.2, creating or altering a namespace with a custom location outside its parent location is now prohibited by default. To restore the old behavior, you can set the ALLOW_NAMESPACE_CUSTOM_LOCATION flag to true.

```

We used to set the namepsace location by ourselves. Instead, let's rely on Polaris, which by default uses:
```
<warehouse>/<namespace>/<table>/
```

And, that's what we look for anyway. The `warehouse` is `database name` for our purposes.